Added check exist email and changed database

+++ database/migrations/2020_06_20_042908_create_books_table.php
@@ -23,7 +23,7 @@ class CreateBooksTable extends Migration
             $table->string('editor', 50)->nullable();
             $table->text('description')->nullable();
             $table->bigInteger('user_id');
-            $table->integer('rating')->default(0)->unsigned();
+            $table->double('rating', 8, 2)->default(0)->unsigned();
             $table->integer('comment_count')->default(0)->unsigned();
             $table->timestamps();
             $table->softDeletes();
